Move PURL-related Identifier functions to Extensions
To have them close to the related ClearlyDefined mapping functions, and
to keep the Identifier class more lightweight.
