+++ model/src/main/kotlin/utils/Extensions.kt
@@ -24,9 +24,11 @@ import org.ossreviewtoolkit.clearlydefined.ClearlyDefinedService.SourceLocation
 import org.ossreviewtoolkit.clearlydefined.ComponentType
 import org.ossreviewtoolkit.clearlydefined.Provider
 import org.ossreviewtoolkit.model.Identifier
+import org.ossreviewtoolkit.model.Package
 import org.ossreviewtoolkit.model.RemoteArtifact
 import org.ossreviewtoolkit.model.TextLocation
 import org.ossreviewtoolkit.model.VcsInfoCurationData
+import org.ossreviewtoolkit.utils.percentEncode
 
 internal fun TextLocation.prependPath(prefix: String): String =
     if (prefix.isBlank()) path else "${prefix.removeSuffix("/")}/$path"
@@ -112,3 +114,72 @@ fun Identifier.toClearlyDefinedSourceLocation(
         else -> null
     }
 }
+
+enum class PurlType(private val value: String) {
+    ALPINE("alpine"),
+    A_NAME("a-name"),
+    BOWER("bower"),
+    CARGO("cargo"),
+    COCOAPODS("cocoapods"),
+    COMPOSER("composer"),
+    CONAN("conan"),
+    CONDA("conda"),
+    CRAN("cran"),
+    DEBIAN("debian"),
+    DRUPAL("drupal"),
+    GEM("gem"),
+    GOLANG("golang"),
+    MAVEN("maven"),
+    NPM("npm"),
+    NUGET("nuget"),
+    PECOFF("pecoff"),
+    PYPI("pypi"),
+    RPM("rpm");
+
+    override fun toString() = value
+}
+
+/**
+ * Map a [Package]'s type to the string representation of the respective [PurlType], or fall back to the lower-case
+ * [Package]'s type if the [PurlType] cannot be determined.
+ */
+fun Identifier.getPurlType() =
+    when (val lowerType = type.toLowerCase()) {
+        "bower" -> PurlType.BOWER
+        "composer" -> PurlType.COMPOSER
+        "conan" -> PurlType.CONAN
+        "crate" -> PurlType.CARGO
+        "godep", "gomod" -> PurlType.GOLANG
+        "gem" -> PurlType.GEM
+        "maven" -> PurlType.MAVEN
+        "npm" -> PurlType.NPM
+        "nuget" -> PurlType.NUGET
+        "pypi" -> PurlType.PYPI
+        else -> lowerType
+    }.toString()
+
+/**
+ * Create the canonical [package URL](https://github.com/package-url/purl-spec) ("purl") based on the properties of
+ * the [Identifier]. Some issues remain with this specification
+ * (see e.g. https://github.com/package-url/purl-spec/issues/33).
+ *
+ * This implementation uses the package type as 'type' purl element as it is used
+ * [in the documentation](https://github.com/package-url/purl-spec/blob/master/README.rst#purl).
+ * E.g. 'maven' for Gradle projects.
+ */
+fun Identifier.toPurl() = "".takeIf { this == Identifier.EMPTY }
+    ?: buildString {
+        append("pkg:")
+        append(getPurlType())
+
+        if (namespace.isNotEmpty()) {
+            append('/')
+            append(namespace.percentEncode())
+        }
+
+        append('/')
+        append(name.percentEncode())
+
+        append('@')
+        append(version.percentEncode())
+    }
